support more cpufreq drivers
The acpi-cpufreq doesn't have policy directories. However, all of the
drivers seem to have cpuN/cpufreq directories, so use those instead.
Read freqdomain_cpus if present to figure out which CPUs are in the
same domain.

Also read and use scaling_available_frequencies for drivers that
provide a discrete set of speeds.